refactor(ui): wire TopologyPage to the existing topology/ helpers
A previous refactor attempt split out ui/src/pages/topology/{DeviceNode,
DeviceDetailsPanel, TopologyLegend, layout, types}.tsx but never
updated TopologyPage.tsx to actually use them — the page kept inline
copies, which had drifted (PR #524 added the edges==0 grid fallback
to the inline layoutNodes but not the topology/ one). 821 lines of
the page was duplicated helpers.

Reconcile by deleting the inline duplicates and importing through the
existing barrel. Also ports the grid-fallback fix into the canonical
topology/layout.ts so the behaviour from #524 survives.

TopologyPage.tsx now 821 -> 392, comfortably under the 500-line red
flag. Helpers themselves stay where they were:

  topology/types.ts (54)             — DeviceNodeData, LinkEdgeData,
                                       DeviceNode, LinkEdge,
                                       linkSpeedColors
  topology/layout.ts (229)           — layoutNodes (with grid fallback)
                                       + edge styling helpers
                                       + createEdges
  topology/DeviceNode.tsx (100)      — custom ReactFlow node
  topology/TopologyLegend.tsx (98)   — collapsible legend
  topology/DeviceDetailsPanel.tsx (94)— floating selected-device card
  topology/NeighborDiscoveryTable.tsx (82) — present but unused; the
                                       page still inlines the table

Pure code consolidation; behaviour unchanged. biome / tsc / vitest /
make build all pass.
